pyproject.toml
[project]
name = "credipred"
version = "0.1.0"
description = "Temporal Graph Analysis on Large-Scale Web Data"
readme = "README.md"
requires-python = ">=3.9"
authors = [
]
dependencies = [
    "boto3>=1.38.33",
    "hf-argparser>=0.1.1",
    "idna>=3.10",
    "mypy>=1.16.1",
    "orjson>=3.10.18",
    "pyspark>=4.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "sentence-transformers>=5.0.0",
    "tldextract>=5.3.0",
    "torch>=2.7.1",
    "torch-geometric>=2.6.1",
    "torcheval>=0.0.7",
    "ujson>=5.10.0",
    "warcio>=1.7.5",
    "cloudscraper>=1.2.71",
    "BeautifulSoup4>=4.13.0",
    "jsonpath-ng>=1.7.0",
    "wandb>=0.18.0",
]

[tool.flit.module]
name = "credipred"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "matplotlib>=3.9.4",
    "ogb>=1.3.6",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "pyqt6>=6.9.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "types-pyyaml>=6.0.12.20240917",
]
analytics = [
]

[tool.mypy]
files = ['credipred']
exclude = ['^credipred/cc-scripts/', '^test/']
install_types = true
non_interactive = true
ignore_missing_imports = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
skip = [".gitignore", "__init__.py"]

[tool.ruff]
exclude = ['test']

[tool.ruff.lint]
select = [
    "D",  # pydocstyle
]
ignore = [
    "D100",  # Ignore "Missing docstring in public module"
    "D101",  # Ignore "Missing docstring in public class"
    "D102",  # Ignore "Missing docstring in public method"
    "D103",  # Ignore "Missing docstring in public function"
    "D104",  # Ignore "Missing docstring in public package"
    "D105",  # Ignore "Missing docstring in magic method"
    "D107",  # Ignore "Missing docstring in __init__"
    "D205",  # Ignore "blank line required between summary line and description"
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
addopts = [
    "--capture=no",
    "--color=yes",
    "-vvv",
]
markers = [
    "integration: marks integration tests (deselect with '-m \"not integration\"')"
]


[tool.coverage.run]
source = ["credipred"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pass",
    "raise NotImplementedError",
    "register_parameter",
    "torch.cuda.is_available",
]
